feat: add dino colors; refactor and update repository/code structure
