+++ chrome_trex/__init__.py
@@ -1,4 +1,4 @@
-from chrome_trex.dinogame import (  # noqa: F401
+from chrome_trex.constants import (  # noqa: F401
     ACTION_DOWN,
     ACTION_FORWARD,
     ACTION_UP,
@@ -7,14 +7,18 @@ from chrome_trex.dinogame import (  # noqa: F401
     HEIGHT,
     WHITE,
     WIDTH,
+)
+from chrome_trex.core.dino_game import DinoGame  # noqa: F401
+from chrome_trex.core.multi_dino_game import (  # noqa: F401
     Cactus,
     Cloud,
     Dino,
-    DinoGame,
     Ground,
     MultiDinoGame,
     Ptera,
     Scoreboard,
+)
+from chrome_trex.helpers import (  # noqa: F401
     extract_digits,
     load_image,
     load_sprite_sheet,
