+++ chrome_trex/constants.py
@@ -1,0 +1,11 @@
+ACTION_FORWARD = 0
+ACTION_UP = 1
+ACTION_DOWN = 2
+
+BLACK = (0, 0, 0)
+WHITE = (255, 255, 255)
+BACKGROUND_COL = WHITE
+
+WIDTH, HEIGHT = 600, 150
+
+SPRITE_SCALE_Y = int(11 * 6 / 5)
